Reslove Jenkins build error
Change readFValsAsInt and readAPValsAsInt arguemnts from int to int64

+++ src/global_det_data.F90
@@ -517,7 +517,7 @@ contains
     subroutine readFValsAsInt(fvals, j)
       implicit none
       integer(n_int), intent(in) :: fvals(:)
-      integer, intent(in) :: j
+      integer(int64), intent(in) :: j
 
       integer :: run
       real(dp) :: realVal = 0.0_dp
@@ -718,7 +718,7 @@ contains
     subroutine readAPValsAsInt(apvals, j)
       implicit none
       integer(n_int), intent(in) :: apvals(:)
-      integer, intent(in) :: j
+      integer(int64), intent(in) :: j
 
       integer :: k
       real(dp) :: realVal = 0.0_dp
